refactor(components): migrate ModalHeader to TypeScript
feature/modal-hook
